+++ src/test/java/http/ProtocolTest.java
@@ -1,5 +1,6 @@
-package http.request;
+package http;
 
+import http.request.protocol.Protocol;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
