+++ src/test/java/webserver/response/HttpStatusCodeTest.java
@@ -1,5 +1,6 @@
 package webserver.response;
 
+import http.response.HttpStatusCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
